feat: Refactor Visualizer.

+++ vis4d/vis/util.py
@@ -6,20 +6,20 @@ import colorsys
 import numpy as np
 
 
-def generate_color_map(length: int) -> list[tuple[float, float, float]]:
+def generate_color_map(length: int) -> list[tuple[int, int, int]]:
     """Generate a color palette of [length] colors.
 
     Args:
         length (int): Number of colors to generate.
 
     Returns:
-        list[tuple[float, float, float]]: List with different colors ranging
+        list[tuple[int, int, int]]: List with different colors ranging
             from [0,255].
     """
     brightness = 0.7
     hsv = [(i / length, 1, brightness) for i in range(length)]
     colors_float = [colorsys.hsv_to_rgb(*c) for c in hsv]
-    colors: list[float] = (
+    colors: list[int] = (
         (np.array(colors_float) * 255).astype(np.uint8).tolist()
     )
     s = np.random.get_state()
